requirements.txt
openai
streamlit
chromadb
tiktoken
python-dotenv
firecrawl-py
tqdm
watchdog